pyproject.toml
[tool.poetry]
name = "cellgenerator"
version = "0.1.0"
description = ""
authors = ["Hugh Warden <hugh.warden@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
numpy = "^2.1.2"
scipy = "^1.14.1"
pillow = "^11.0.0"
matplotlib = "^3.9.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
